# doc/restapi/restapi.swagger2.json
{
  "swagger": "2.0",
  "info": {
    "version": "v1",
    "title": "API V1"
  },
  "basePath": "/",
  "paths": {
...
    "/api/Group/{id}": {
      "delete": {
        "tags": [ "Group" ],
        "summary": "Delete group with the <param name=\"id\" /> provided.\r\n            \r\n            Example Usage: DELETE api/group/1",
        "operationId": "ApiGroupByIdDelete",
        "consumes": [],
        "produces": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Group ID.",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": { "204": { "description": "No Content" } },
        "deprecated": false
      }
    },
    "/api/GroupMember/requests/{groupId}": {
      "get": {
        "tags": [ "GroupMember" ],
        "summary": "Get a list of all Users that have relationship requests for this <param name=\"groupId\" />.\r\n            \r\n            Example Usage: GET api/groupmember/requests/1",
        "operationId": "ApiGroupMemberRequestsByGroupIdGet",
        "consumes": [],
        "produces": [],
        "parameters": [
          {
            "name": "groupId",
            "in": "path",
            "description": "ID of the group.",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": { "204": { "description": "No Content" } },
        "deprecated": false
      }
    },
    "/api/GroupMember/sentrequests/{userId}": {
      "get": {
        "tags": [ "GroupMember" ],
        "summary": "Get a list of all Groups that have been sent relationship requests for this <param name=\"userId\" />.\r\n            \r\n            Example Usage: GET api/groupmember/sentrequests/1",
        "operationId": "ApiGroupMemberSentrequestsByUserIdGet",
        "consumes": [],
        "produces": [],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "ID of the user.",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": { "204": { "description": "No Content" } },
        "deprecated": false
      }
    },
    "/api/GroupMember/members/{groupId}": {
      "get": {
        "tags": [ "GroupMember" ],
        "summary": "Get a list of all Users that have relationships with this <param name=\"groupId\" />.\r\n            \r\n            Example Usage: GET api/groupmember/members/1",
        "operationId": "ApiGroupMemberMembersByGroupIdGet",
        "consumes": [],
        "produces": [],
        "parameters": [
          {
            "name": "groupId",
            "in": "path",
            "description": "ID of the group.",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": { "204": { "description": "No Content" } },
        "deprecated": false
      }
    },
    "/api/GroupMember/usergroups/{userId}": {
      "get": {
        "tags": [ "GroupMember" ],
        "summary": "Get a list of all Groups that have relationships with this <param name=\"userId\" />.\r\n            \r\n            Example Usage: GET api/groupmember/usergroups/1",
        "operationId": "ApiGroupMemberUsergroupsByUserIdGet",
        "consumes": [],
        "produces": [],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "ID of the User.",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": { "204": { "description": "No Content" } },
        "deprecated": false
      }
    },
    "/api/GroupMember": {
      "put": {
        "tags": [ "GroupMember" ],
        "summary": "Update an existing relationship between <param name=\"relationship.UserId.UserId\" /> and <param name=\"relationship.GroupId.GroupId\" />.\r\n            Requires the relationship to already exist between the User and Group.\r\n            \r\n            Example Usage: PUT api/groupmember",
        "operationId": "ApiGroupMemberPut",
        "consumes": [ "application/json", "text/json", "application/json-patch+json" ],
        "produces": [],
        "parameters": [
          {
            "name": "relationship",
            "in": "body",
            "description": "{PlayGen.SUGAR.Contracts.RelationshipStatusUpdate} object that holds the details of the relationship.",
            "required": false,
            "schema": { "$ref": "#/definitions/RelationshipStatusUpdate" }
          }
        ],
        "responses": { "204": { "description": "No Content" } },
        "deprecated": false
      },
      "post": {
        "tags": [ "GroupMember" ],
        "summary": "Create a new relationship request between the User and Group.\r\n            Requires a relationship between the User and Group to not already exist.\r\n            \r\n            Example Usage: POST api/groupmember",
        "operationId": "ApiGroupMemberPost",
        "consumes": [ "application/json", "text/json", "application/json-patch+json" ],
        "produces": [],
        "parameters": [
          {
            "name": "relationship",
            "in": "body",
            "description": "{PlayGen.SUGAR.Contracts.RelationshipRequest} object that holds the details of the new relationship request.",
            "required": false,
            "schema": { "$ref": "#/definitions/RelationshipRequest" }
          }
        ],
        "responses": { "204": { "description": "No Content" } },
        "deprecated": false
      }
    },
    "/api/GroupMember/request": {
      "put": {
        "tags": [ "GroupMember" ],
        "summary": "Update an existing relationship request between <param name=\"relationship.UserId.UserId\" /> and <param name=\"relationship.GroupId.GroupId\" />.\r\n            Requires the relationship request to already exist between the User and Group.\r\n            \r\n            Example Usage: PUT api/groupmember/request",
        "operationId": "ApiGroupMemberRequestPut",
        "consumes": [ "application/json", "text/json", "application/json-patch+json" ],
        "produces": [],
        "parameters": [
          {
            "name": "relationship",
            "in": "body",
            "description": "{PlayGen.SUGAR.Contracts.RelationshipStatusUpdate} object that holds the details of the relationship.",
            "required": false,
            "schema": { "$ref": "#/definitions/RelationshipStatusUpdate" }
          }
        ],
        "responses": { "204": { "description": "No Content" } },
        "deprecated": false
      }
    },
...
  },
  "definitions": {
...
    "RelationshipRequest": {
      "description": "Encapsulates relationship details.",
      "required": [ "requestorId", "acceptorId", "autoAccept" ],
      "type": "object",
      "properties": {
        "requestorId": {
          "format": "int32",
          "description": "The Id of the requesting actor.",
          "type": "integer"
        },
        "acceptorId": {
          "format": "int32",
          "description": "The Id of the receiving actor.",
          "type": "integer"
        },
        "autoAccept": {
          "description": "Whether the request should be automatically accepted.",
          "type": "boolean"
        }
      }
    },
    "RelationshipStatusUpdate": {
      "description": "Encapsulates relationship details, including updated status of the relationship.",
      "required": [ "requestorId", "acceptorId", "accepted" ],
      "type": "object",
      "properties": {
        "requestorId": {
          "format": "int32",
          "description": "The Id of the requesting actor.",
          "type": "integer"
        },
        "acceptorId": {
          "format": "int32",
          "description": "The Id of the receiving actor.",
          "type": "integer"
        },
        "accepted": {
          "description": "Whether the request was accepted or declined.",
          "type": "boolean"
        }
      }
    },
...
  },
  "securityDefinitions": {}
}